+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## [Unreleased]
 
+- Add `timezone` to variables ([Issue #69](https://github.com/open-meteo-ruby/open-meteo-ruby/issues/69))
 - fix(logging): uses configured logger with Faraday https://github.com/cpb
 
 ## [0.3.2] - 2024-03-21